ovn-controller: Avoid recompute triggered by external_ids:ovn-installed update.
During the process of claiming a VIF port on a chassis, there are
multiple SB and local OVSDB updates by ovn-controller, which may trigger
ovn-controller recompute, depending on the timing between the completion
of SB and local OVSDB updates:

After port claiming, ovn-controller sends an update to the local OVSDB
to clean the Interface:external_ids:ovn-installed, and also sends an
update to the SB DB to update the Port_Binding:chassis.  If the local
OVSDB update notification comes back before the SB update completes,
ovn-controller's SB IDL is still "in transaction" and read-only, but the
local Interface update triggers the I-P engine runtime_data node to
handle the OVS_interface change, which considers the VIF to be claimed
(again) potentially, which requires SB updates while it is not possible
at the moment, and so the I-P handler returns failure and the engine
aborts, which triggers recompute at the next I-P engine iteration when
SB DB is writable.

If we are lucky at the above step that the SB update completes before
the local OVSDB update is received, ovn-controller would handle the
change incrementally (and figures out that the SB Port_Binding chassis
is already the current one thus wouldn't reclaim the VIF). But the next
step after it completes flow computing for the newly bound VIF it sends
another update to the local OVSDB to set the
Interface:external_ids:ovn-installed and ovn-installed-ts, while at the
same time it also sends an update to the SB DB to set the
Port_Binding:up. These two parallel updates would lead to the same
situation as the above step and potentially cause ovn-controll
recompute.

The recompute is very likely to be triggered in production because the
SB DB is remote (on central nodes) and has heavier load, which requires
longer time to complete the updates than the local OVSDB server.

The problem is that the runtime_data's OVS_interface change handler
couldn't handle the update of the Interface:external_ids:ovn-installed
incrementally, because ovn-controller assumes it could require an SB
update that is not possible at the moment. However, the assumption is
not true. In fact it shouldn't even care about the change because this
external_id:ovn-installed is written by ovn-controller itself and it is
write-only (no need to read it as an input for any further processing).

So, the idea to fix the problem is to ignore such changes that happen to
the "write-only" fields only. OVSDB change tracking provides APIs to
check if a specific column is updated, but the difficulty here is that
the column external_ids is a k-v map which contains not only the data
that ovn-controller can ignore, such as ovn-installed, ovn-installed-ts,
but also the data ovn-controller cares about, such as iface-id,
iface-id-ver and encap-ip. To solve this problem, we need to know what's
changed exactly inside the external_ids map.

This patch solves it by introducing a new I-P engine node
ovs_interface_shadow to wrap the OVSDB input ovs_interface with an extra
copy of the older version of the column external_ids, and replace the
input OVS_interface of the runtime_data node. In the change handler of the
runtime_data node we evaluates if the change needs to be handled before
calling consider_iface_claim(), so in the above scenarios the changes
don't need to be handled and won't trigger recompute any more. A test
case is also updated to capture this (which is very likely, if not 100%,
to fail without the fix).

Signed-off-by: Han Zhou <hzhou@ovn.org>