ovn-northd: Avoid recompute caused by in-flight transactions.
Although each individual VIF port related changes are handled
incrementally, it still triggers recompute if there are in-flight
transactions (either to NB or SB) when a change comes, which is very
common in real production environment if change happens frequently.
It is also easy to hit such situation in test cases where nb_cfg
mechanism is heavily used, which makes it difficult to write reliable
and stable tests, such as what the commit 8c30ba1386 was trying to work
around.

This patch skips the I-P engine execution until the NB & SB transaction
handles are available (no in-flight transactions), and when skippiing
the runs it keeps the tracked changes in IDL across main loop
iterations. This way we avoid recompute without worrying about missing
any changes.

Signed-off-by: Han Zhou <hzhou@ovn.org>