physical: Don't reset encap ID across pipelines.
The MFF_LOG_ENCAP_ID register was defined to save the encap ID and avoid
changing across pipelines, but in the function
load_logical_ingress_metadata it was reset unconditionally. Because of
this, the encap selection doesn't work when traffic traverses L3
boundaries. This patch fixes it by ensuring the register is loaded only
in the flows of table 0, which is where packets from VIFs enter the
pipeline for the first time.

Fixes: 17b6a12fa286 ("ovn-controller: Support VIF-based local encap IPs selection.")
Signed-off-by: Han Zhou <hzhou@ovn.org>