northd: Fix HW offload problem related to ct_tuple.
Commit 20a96b9d2c3d used CT header fields for L4 matching, which caused
HW offload problems for LB related traffic.

The root cause is, it added openflow rules with match ct_state=new and
ct_nw_proto=xxx. The original consideration was that, although the
ct_nw_proto is not supported by HW offload, it is only combined with
ct_state=new, so it was assumed that it only affects the first packet of
a new connection, which is not offloaded anyway.  However, that's not
the way how OVS is implemented. OVS generates megaflows not just
according to what's matched in openflow tables, but also what's not
matched. In this case, when a packet of an established connections
comes, although it doesn't match the flow that has "ct_state=new &&
ct_nw_proto=xxx", just because it is examined against the rule, the
megaflow unwildcards the ct_nw_proto field as exact match.  So in the
end the packet matches another flow "ct_state=est ..." that doesn't have
ct_nw_proto in the mtach, but still generates a megaflow with
ct_tuple4(proto=xxx) (which is a result of the ct_nw_proto condition of
the openflow rule). And ct_tuple4... is not supported for HW offload, so
all established connection packets are impacted and not offloaded.

In theory OVS could be optimized so that when ct_state is not matched,
then don't consider other ct fields of the same flow. However, this is
challenging.

So this patch introduces a separate stage CT_EXTRACT for both LS and LR
pipelines, which jumps to a side-table to access the CT fields, only if
the CT state is new. This way, the CT fields won't appear in the match
conditions for established connections, thus won't generate megaflows
with ct_tuple4(proto=xxx).

Fixes: 20a96b9d2c3d ("northd: Allow fragmented traffic through LB.")
Signed-off-by: Han Zhou <hzhou@ovn.org>